requirements.txt
numpy
openai
pdfplumber
python-docx
streamlit
yagmail